+++ kikuchipy/indexing/similarity_metrics.py
@@ -16,7 +16,7 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
-"""Similarity metrics for comparing gray-tone images."""
+"""Similarity metrics for comparing gray-tone patterns."""
 
 from enum import Enum
 from typing import Callable, Tuple, Union
@@ -44,58 +44,59 @@ def make_similarity_metric(
     make_compatible_to_lower_scopes: bool = False,
     dtype_out: np.dtype = np.float32,
 ):
-    """Make a similarity metric for comparing gray-tone images of equal
-    size.
+    """Make a similarity metric for comparing gray-tone patterns of
+    equal size.
 
     This factory function wraps metric functions for use in
-    `template_match`, which again is used by
-    :class:`~kikuchipy.indexing.StaticDictionary` and
-    :class:`~kikuchipy.indexing.DynamicDictionary`.
+    :func:`~kikuchipy.indexing.pattern_matching.pattern_match`, which
+    again is used by
+    :class:`~kikuchipy.indexing.StaticDictionaryIndexing` and
+    :class:`~kikuchipy.indexing.DynamicDictionaryIndexing`.
 
     Parameters
     ----------
-    metric_func
+    metric_func : Callable
         Metric function with signature
-        `metric_func(patterns, templates)`, which computes the
-        similarity or a distance matrix between (experimental)
-        pattern(s) and (simulated) template(s).
-    greater_is_better
-        Whether greater values correspond to more similar images, by
-        default True. Used for choosing `n_largest` metric results in
-        `template_match`.
-    scope
+        `metric_func(experimental, simulated)`, which computes the
+        similarity or a distance matrix between experimental
+        and simulated pattern(s)).
+    greater_is_better : bool, optional
+        Whether greater values correspond to more similar patterns, by
+        default True. Used for choosing `keep_n` metric results in
+        `pattern_match`.
+    scope : MetricScope, optional
         Describes the shapes of the `metric_func`'s input parameters
-        `patterns` and `templates`, by default
+        `experimental` and `simulated`, by default
         `MetricScope.MANY_TO_MANY`.
-    flat
-        Whether patterns and templates are to be flattened before sent
-        to `metric_func` when the similarity metric is called, by
-        default False.
-    make_compatible_to_lower_scopes
-        Whether to reshape patterns and templates by adding single
-        dimensions to match the given scope, by default False.
-    dtype_out
+    flat : bool, optional
+        Whether experimental and simulated patterns are to be flattened
+        before sent to `metric_func` when the similarity metric is
+        called, by default False.
+    make_compatible_to_lower_scopes : bool, optional
+        Whether to reshape experimental and simulated data by adding
+        single dimensions to match the given scope, by default False.
+    dtype_out : numpy.dtype, optional
         The data type used and returned by the metric, by default
-        :class:`np.float32`.
+        :class:`numpy.float32`.
 
     Returns
     -------
     metric_class : SimilarityMetric or FlatSimilarityMetric
         A callable class instance computing a similarity matrix with
-        signature `metric_func(patterns, templates)`.
+        signature `metric_func(experimental, simulated)`.
 
     Notes
     -----
-    The metric function must take the arrays `patterns` and `templates`
-    as arguments, in that order. The scope and whether the metric is
-    flat defines the intended data shapes. In the following table,
-    (ny, nx) and (sy, sx) correspond to the navigation and signal
-    shapes (row, column), respectively, with N number of templates:
+    The metric function must take the arrays `experimental` and
+    `simulated` as arguments, in that order. The scope and whether the
+    metric is flat defines the intended data shapes. In the following
+    table, (ny, nx) and (sy, sx) correspond to the navigation and signal
+    shapes (row, column), respectively, with N number of simulations:
 
     ============ ============= ========= ========= ============= ========= =========
     MetricScope  flat = False                      flat = True
     ------------ --------------------------------- ---------------------------------
-    \-           patterns      templates returns   patterns      templates returns
+    \-           experimental  simulated returns   experimental  simulated returns
     ============ ============= ========= ========= ============= ========= =========
     MANY_TO_MANY (ny,nx,sy,sx) (N,sy,sx) (ny,nx,N) (ny*nx,sy*sx) (N,sy*sx) (ny*nx,N)
     ONE_TO_MANY  (sy,sx)       (N,sy,sx) (N,)      (sy*sx,)      (N,sy*sx) (N,)
@@ -125,18 +126,18 @@ def make_similarity_metric(
 
 
 class SimilarityMetric:
-    """Similarity metric between 2D gray-tone images."""
+    """Similarity metric between 2D gray-tone patterns."""
 
     # See table in docstring of `make_similarity_metric`
     # TODO: Support for 1D navigation shape
-    _P_T_NDIM_TO_SCOPE = {
+    _EXPT_SIM_NDIM_TO_SCOPE = {
         (4, 3): MetricScope.MANY_TO_MANY,
         (2, 3): MetricScope.ONE_TO_MANY,
         (4, 2): MetricScope.MANY_TO_ONE,
         (2, 2): MetricScope.ONE_TO_ONE,
     }
 
-    _SCOPE_TO_P_T_NDIM = {
+    _SCOPE_TO_EXPT_SIM_NDIM = {
         MetricScope.MANY_TO_MANY: (4, 3),
         MetricScope.ONE_TO_MANY: (2, 3),
         MetricScope.MANY_TO_ONE: (4, 2),
@@ -178,21 +179,21 @@ class SimilarityMetric:
 
     def __call__(
         self,
-        patterns: Union[np.ndarray, da.Array],
-        templates: Union[np.ndarray, da.Array],
+        experimental: Union[np.ndarray, da.Array],
+        simulated: Union[np.ndarray, da.Array],
     ) -> Union[np.ndarray, da.Array]:
         dtype = self._dtype_out
-        patterns = patterns.astype(dtype)
-        templates = templates.astype(dtype)
-        if isinstance(patterns, da.Array):
-            patterns = patterns.rechunk()
-        if isinstance(templates, da.Array):
-            templates = templates.rechunk()
+        experimental = experimental.astype(dtype)
+        simulated = simulated.astype(dtype)
+        if isinstance(experimental, da.Array):
+            experimental = experimental.rechunk()
+        if isinstance(simulated, da.Array):
+            simulated = simulated.rechunk()
         if self._make_compatible_to_lower_scopes:
-            patterns, templates = self._expand_dims_to_match_scope(
-                patterns, templates
+            experimental, simulated = self._expand_dims_to_match_scope(
+                experimental, simulated
             )
-        return self._measure(patterns, templates).squeeze()
+        return self._measure(experimental, simulated).squeeze()
 
     def __repr__(self):
         return (
@@ -202,31 +203,37 @@ class SimilarityMetric:
 
     def _measure(
         self,
-        patterns: Union[np.ndarray, da.Array],
-        templates: Union[np.ndarray, da.Array],
+        experimental: Union[np.ndarray, da.Array],
+        simulated: Union[np.ndarray, da.Array],
     ) -> Union[np.ndarray, da.Array]:
         """Measure the similarities and return the results."""
-        return self._metric_func(patterns, templates)
+        return self._metric_func(experimental, simulated)
 
     def _expand_dims_to_match_scope(
-        self, p: Union[np.ndarray, da.Array], t: Union[np.ndarray, da.Array],
+        self,
+        expt: Union[np.ndarray, da.Array],
+        sim: Union[np.ndarray, da.Array],
     ) -> Tuple[Union[np.ndarray, da.Array], Union[np.ndarray, da.Array]]:
-        """Return patterns and templates with added axes corresponding
-        to the scope of the metric.
+        """Return experimental and simulated data with added axes
+        corresponding to the scope of the metric.
         """
-        p_scope_ndim, t_scope_ndim = self._SCOPE_TO_P_T_NDIM[self.scope]
-        p = p[(np.newaxis,) * (p_scope_ndim - p.ndim)]
-        t = t[(np.newaxis,) * (t_scope_ndim - t.ndim)]
-        return p, t
+        expt_scope_ndim, sim_scope_ndim = self._SCOPE_TO_EXPT_SIM_NDIM[
+            self.scope
+        ]
+        expt = expt[(np.newaxis,) * (expt_scope_ndim - expt.ndim)]
+        sim = sim[(np.newaxis,) * (sim_scope_ndim - sim.ndim)]
+        return expt, sim
 
-    def _is_compatible(self, p_ndim: int, t_ndim: int) -> bool:
-        """Return whether shapes of patterns and templates are
+    def _is_compatible(self, expt_ndim: int, sim_ndim: int) -> bool:
+        """Return whether shapes of experimental and simulated are
         compatible with the metric's scope.
         """
         if self.flat:
-            p_ndim = p_ndim // 2  # 4 -> 2 or 2 -> 1
-            t_ndim -= 1
-        inferred_scope = self._P_T_NDIM_TO_SCOPE.get((p_ndim, t_ndim), False)
+            expt_ndim = expt_ndim // 2  # 4 -> 2 or 2 -> 1
+            sim_ndim -= 1
+        inferred_scope = self._EXPT_SIM_NDIM_TO_SCOPE.get(
+            (expt_ndim, sim_ndim), False
+        )
         if not inferred_scope:
             return False
         if inferred_scope == self.scope:
@@ -244,14 +251,14 @@ class FlatSimilarityMetric(SimilarityMetric):
     """
 
     # See table in docstring of `make_similarity_metric`
-    _P_T_NDIM_TO_SCOPE = {
+    _EXPT_SIM_NDIM_TO_SCOPE = {
         (2, 2): MetricScope.MANY_TO_MANY,
         (1, 2): MetricScope.ONE_TO_MANY,
         (3, 1): MetricScope.MANY_TO_ONE,
         (1, 1): MetricScope.ONE_TO_ONE,
     }
 
-    _SCOPE_TO_P_T_NDIM = {
+    _SCOPE_TO_EXPT_SIM_NDIM = {
         MetricScope.MANY_TO_MANY: (2, 2),
         MetricScope.ONE_TO_MANY: (1, 2),
         MetricScope.MANY_TO_ONE: (3, 1),
@@ -259,204 +266,211 @@ class FlatSimilarityMetric(SimilarityMetric):
     }
 
     _PARTIAL_SCOPE_TO_NDIM = {
-        "MANY_": 2,  # Many patterns
-        "_MANY": 2,  # Many templates
+        "MANY_": 2,  # Many experimental
+        "_MANY": 2,  # Many simulated
         "ONE_": 1,  # One pattern
         "_ONE": 1,  # One template
     }
 
     def _measure(
         self,
-        patterns: Union[np.ndarray, da.Array],
-        templates: Union[np.ndarray, da.Array],
+        experimental: Union[np.ndarray, da.Array],
+        simulated: Union[np.ndarray, da.Array],
     ) -> Union[np.ndarray, da.Array]:
         """Flatten images, measure the similarities and return the
         results.
         """
-        nav_shape = _get_nav_shape(patterns)
-        sig_shape = _get_sig_shape(patterns)
+        nav_shape = _get_nav_shape(experimental)
+        sig_shape = _get_sig_shape(experimental)
         nav_flat_size, sig_flat_size = np.prod(nav_shape), np.prod(sig_shape)
-        patterns = patterns.reshape((nav_flat_size, sig_flat_size))
-        templates = templates.reshape((-1, sig_flat_size))
-        return self._metric_func(patterns, templates)
+        experimental = experimental.reshape((nav_flat_size, sig_flat_size))
+        simulated = simulated.reshape((-1, sig_flat_size))
+        return self._metric_func(experimental, simulated)
 
 
-def _get_nav_shape(p):
-    return {2: (), 3: (p.shape[0],), 4: (p.shape[:2])}[p.ndim]
+def _get_nav_shape(expt):
+    return {2: (), 3: (expt.shape[0],), 4: (expt.shape[:2])}[expt.ndim]
 
 
-def _get_sig_shape(p):
-    return p.shape[-2:]
+def _get_sig_shape(expt):
+    return expt.shape[-2:]
 
 
-def _get_number_of_templates(t):
-    return t.shape[0] if t.ndim == 3 else 1
+def _get_number_of_simulated(sim):
+    return sim.shape[0] if sim.ndim == 3 else 1
 
 
 def _expand_dims_to_many_to_many(
-    p: Union[np.ndarray, da.Array], t: Union[np.ndarray, da.Array], flat: bool,
+    expt: Union[np.ndarray, da.Array],
+    sim: Union[np.ndarray, da.Array],
+    flat: bool,
 ) -> Tuple[Union[np.ndarray, da.Array], Union[np.ndarray, da.Array]]:
-    """Expand the dims of patterns and templates to match
+    """Expand the dims of experimental and simulated to match
     `MetricScope.MANY_TO_MANY`.
 
     Parameters
     ----------
-    p
-        Patterns.
-    t
-        Templates.
+    expt
+        Experimental data.
+    sim
+        Simulated data.
     flat
-        Whether `p` and `t` are flattened.
+        Whether `expt` and `sim` are flattened.
 
     Returns
     -------
-    p
-        Patterns with their dimension expanded to match
+    expt
+        Experimental data dimension expanded to match
         `MetricScope.MANY_TO_MANY`.
-    t
-        Templates with their dimension expanded to match
+    sim
+        Simulated data dimension expanded to match
         `MetricScope.MANY_TO_MANY`.
     """
     metric_cls = FlatSimilarityMetric if flat else SimilarityMetric
-    p_scope_ndim, t_scope_ndim = metric_cls._SCOPE_TO_P_T_NDIM[
+    expt_scope_ndim, sim_scope_ndim = metric_cls._SCOPE_TO_EXPT_SIM_NDIM[
         MetricScope.MANY_TO_MANY
     ]
-    p = p[(np.newaxis,) * (p_scope_ndim - p.ndim)]
-    t = t[(np.newaxis,) * (t_scope_ndim - t.ndim)]
-    return p, t
+    expt = expt[(np.newaxis,) * (expt_scope_ndim - expt.ndim)]
+    sim = sim[(np.newaxis,) * (sim_scope_ndim - sim.ndim)]
+    return expt, sim
 
 
 def _zero_mean(
-    p: Union[np.ndarray, da.Array],
-    t: Union[np.ndarray, da.Array],
+    expt: Union[np.ndarray, da.Array],
+    sim: Union[np.ndarray, da.Array],
     flat: bool = False,
 ) -> Tuple[Union[np.ndarray, da.Array], Union[np.ndarray, da.Array]]:
-    """Subtract the mean from patterns and templates of any scope.
+    """Subtract the mean from experimental and simulated data of any
+    scope.
 
     Parameters
     ----------
-    p : np.ndarray or da.Array
-        Patterns.
-    t : np.ndarray or da.Array
-        Templates.
+    expt : np.ndarray or da.Array
+        Experimental.
+    sim : np.ndarray or da.Array
+        Simulated.
     flat : bool, optional
         Whether `p` and `t` are flattened, by default False.
 
     Returns
     -------
-    p
-        Patterns with their mean subtracted.
-    t
-        Templates with their mean subtracted.
+    expt
+        Experimental data with mean subtracted.
+    sim
+        Simulated data with mean subtracted.
     """
-    squeeze = 1 not in p.shape + t.shape
-    p, t = _expand_dims_to_many_to_many(p, t, flat)
-    p_mean_axis = 1 if flat else (2, 3)
-    t_mean_axis = 1 if flat else (1, 2)
-    p -= p.mean(axis=p_mean_axis, keepdims=True)
-    t -= t.mean(axis=t_mean_axis, keepdims=True)
+    squeeze = 1 not in expt.shape + sim.shape
+    expt, sim = _expand_dims_to_many_to_many(expt, sim, flat)
+    expt_mean_axis = 1 if flat else (2, 3)
+    sim_mean_axis = 1 if flat else (1, 2)
+    expt -= expt.mean(axis=expt_mean_axis, keepdims=True)
+    sim -= sim.mean(axis=sim_mean_axis, keepdims=True)
 
     if squeeze:
-        return p.squeeze(), t.squeeze()
+        return expt.squeeze(), sim.squeeze()
     else:
-        return p, t
+        return expt, sim
 
 
 def _normalize(
-    p: Union[np.ndarray, da.Array],
-    t: Union[np.ndarray, da.Array],
+    expt: Union[np.ndarray, da.Array],
+    sim: Union[np.ndarray, da.Array],
     flat: bool = False,
 ) -> Tuple[Union[np.ndarray, da.Array], Union[np.ndarray, da.Array]]:
-    """Normalize patterns and templates of any scope.
+    """Normalize experimental and simulated patterns of any scope.
 
     Parameters
     ----------
-    p : np.ndarray or da.Array
-        Patterns.
-    t : np.ndarray or da.Array
-        Templates.
+    expt : np.ndarray or da.Array
+        Experimental patterns.
+    sim : np.ndarray or da.Array
+        Simulated patterns.
     flat : bool, optional
-        Whether `p` and `t` are flattened, by default False.
+        Whether `expt` and `sim` are flattened, by default False.
 
     Returns
     -------
-    p
-        Patterns divided by their L2 norms.
-    t
-        Templates divided by their L2 norms.
+    expt
+        Experimental patterns divided by their L2 norms.
+    sim
+        Simulated patterns divided by their L2 norms.
     """
-    squeeze = 1 not in p.shape + t.shape
-    p, t = _expand_dims_to_many_to_many(p, t, flat)
-    p_sum_axis = 1 if flat else (2, 3)
-    t_sum_axis = 1 if flat else (1, 2)
-    p /= (p ** 2).sum(axis=p_sum_axis, keepdims=True) ** 0.5
-    t /= (t ** 2).sum(axis=t_sum_axis, keepdims=True) ** 0.5
+    squeeze = 1 not in expt.shape + sim.shape
+    expt, sim = _expand_dims_to_many_to_many(expt, sim, flat)
+    expt_sum_axis = 1 if flat else (2, 3)
+    sim_sum_axis = 1 if flat else (1, 2)
+    expt /= (expt ** 2).sum(axis=expt_sum_axis, keepdims=True) ** 0.5
+    sim /= (sim ** 2).sum(axis=sim_sum_axis, keepdims=True) ** 0.5
 
     if squeeze:
-        return p.squeeze(), t.squeeze()
+        return expt.squeeze(), sim.squeeze()
     else:
-        return p, t
+        return expt, sim
 
 
 def _zncc_einsum(
-    patterns: Union[da.Array, np.ndarray],
-    templates: Union[da.Array, np.ndarray],
+    experimental: Union[da.Array, np.ndarray],
+    simulated: Union[da.Array, np.ndarray],
 ) -> Union[np.ndarray, da.Array]:
     """Compute (lazily) the zero-mean normalized cross-correlation
-    coefficient between patterns and templates.
+    coefficient between experimental and simulated patterns.
 
     Parameters
     ----------
-    patterns
-        Experimental patterns to compare to the templates.
-    templates
-        Templates, typically simulated patterns.
+    experimental
+        Experimental patterns.
+    simulated
+        Simulated patterns.
 
     Returns
     -------
     zncc
         Correlation coefficients in range [-1, 1] for all comparisons,
-        as :class:`np.ndarray` if both `patterns` and `templates` are
-        :class:`np.ndarray`, else :class:`da.Array`.
+        as :class:`np.ndarray` if both `experimental` and `simulated`
+        are :class:`np.ndarray`, else :class:`da.Array`.
 
     Notes
     -----
     Equivalent results are obtained with :func:`dask.Array.tensordot`
     with the `axes` argument `axes=([2, 3], [1, 2]))`.
     """
-    patterns, templates = _zero_mean(patterns, templates)
-    patterns, templates = _normalize(patterns, templates)
-    zncc = da.einsum("ijkl,mkl->ijm", patterns, templates, optimize=True)
-    if isinstance(patterns, np.ndarray) and isinstance(templates, np.ndarray):
+    experimental, simulated = _zero_mean(experimental, simulated)
+    experimental, simulated = _normalize(experimental, simulated)
+    zncc = da.einsum("ijkl,mkl->ijm", experimental, simulated, optimize=True)
+    if isinstance(experimental, np.ndarray) and isinstance(
+        simulated, np.ndarray
+    ):
         return zncc.compute()
     else:
         return zncc
 
 
 def _ndp_einsum(
-    patterns: Union[da.Array, np.ndarray],
-    templates: Union[da.Array, np.ndarray],
+    experimental: Union[da.Array, np.ndarray],
+    simulated: Union[da.Array, np.ndarray],
 ) -> Union[np.ndarray, da.Array]:
-    """Compute the normalized dot product between patterns and
-    templates.
+    """Compute the normalized dot product between experimental and
+    simulated patterns.
 
     Parameters
     ----------
-    patterns
-        Experimental patterns to compare to the templates.
-    templates
-        Templates, typically simulated patterns.
+    experimental
+        Experimental patterns.
+    simulated
+        Simulated patterns.
 
     Returns
     -------
     ndp
         Normalized dot products in range [0, 1] for all comparisons,
-        as :class:`np.ndarray` if both `patterns` and `templates` are
-        :class:`np.ndarray`, else :class:`da.Array`.
+        as :class:`np.ndarray` if both `experimental` and `simulated`
+        are :class:`np.ndarray`, else :class:`da.Array`.
     """
-    patterns, templates = _normalize(patterns, templates)
-    ndp = da.einsum("ijkl,mkl->ijm", patterns, templates, optimize=True)
-    if isinstance(patterns, np.ndarray) and isinstance(templates, np.ndarray):
+    experimental, simulated = _normalize(experimental, simulated)
+    ndp = da.einsum("ijkl,mkl->ijm", experimental, simulated, optimize=True)
+    if isinstance(experimental, np.ndarray) and isinstance(
+        simulated, np.ndarray
+    ):
         return ndp.compute()
     else:
         return ndp