+++ kikuchipy/indexing/tests/test_similarity_metrics.py
@@ -27,7 +27,7 @@ from kikuchipy.indexing.similarity_metrics import (
     MetricScope,
     FlatSimilarityMetric,
     SIMILARITY_METRICS,
-    _get_number_of_templates,
+    _get_number_of_simulated,
 )
 
 
@@ -40,7 +40,7 @@ class TestSimilarityMetrics:
         assert (
             type(
                 make_similarity_metric(
-                    lambda p, t: np.zeros((2, 4))
+                    lambda expt, sim: np.zeros((2, 4))
                     if flat
                     else np.zeros((2, 2, 2)),
                     flat=flat,
@@ -52,81 +52,89 @@ class TestSimilarityMetrics:
 
     def test_zncc(self):
         zncc_metric = SIMILARITY_METRICS["zncc"]
-        # Four patterns
-        p = np.array(
+        # Four experimental data
+        expt = np.array(
             [
                 [[[1, 2], [3, 4]], [[5, 6], [7, 8]]],
                 [[[9, 8], [1, 7]], [[5, 2], [2, 7]]],
             ],
             np.int8,
         )
-        p_da = da.from_array(p)
+        expt_da = da.from_array(expt)
 
         # One perfect match, at [1,0,1] in results, and one close match
-        # Two templates
-        t = np.array([[[5, 3], [2, 7]], [[9, 8], [1, 7]]], np.int8)
-        t_da = da.from_array(t)
+        # Two simulated
+        sim = np.array([[[5, 3], [2, 7]], [[9, 8], [1, 7]]], np.int8)
+        sim_da = da.from_array(sim)
 
         # many to many
-        assert pytest.approx(zncc_metric(p_da, t_da).compute()[1, 0, 1]) == 1
+        assert (
+            pytest.approx(zncc_metric(expt_da, sim_da).compute()[1, 0, 1]) == 1
+        )
 
         # Working with lower scopes, here one to many:
-        assert pytest.approx(zncc_metric(p_da[1, 0], t_da).compute()[1]) == 1
+        assert (
+            pytest.approx(zncc_metric(expt_da[1, 0], sim_da).compute()[1]) == 1
+        )
 
     def test_ndp(self):
         ndp_metric = SIMILARITY_METRICS["ndp"]
-        p = np.array(
+        expt = np.array(
             [
                 [[[1, 2], [3, 4]], [[5, 6], [7, 8]]],
                 [[[9, 8], [1, 7]], [[5, 2], [2, 7]]],
             ],
             np.int8,
         )
-        p_da = da.from_array(p)
+        expt_da = da.from_array(expt)
 
         # One perfect match and one close match
-        t = np.array([[[5, 3], [2, 7]], [[9, 8], [1, 7]]], np.int8)
-        t_da = da.from_array(t)
+        sim = np.array([[[5, 3], [2, 7]], [[9, 8], [1, 7]]], np.int8)
+        sim_da = da.from_array(sim)
 
         # many to many
-        assert pytest.approx(ndp_metric(p_da, t_da).compute()[1, 0, 1]) == 1
+        assert (
+            pytest.approx(ndp_metric(expt_da, sim_da).compute()[1, 0, 1]) == 1
+        )
 
     @pytest.mark.parametrize("metric", ["zncc", "ndp"])
     def test_zncc_ndp_returns_desired_array_type(self, metric):
         metric = SIMILARITY_METRICS[metric]
-        p = np.array(
+        expt = np.array(
             [
                 [[[1, 2], [3, 4]], [[5, 6], [7, 8]]],
                 [[[9, 8], [1, 7]], [[5, 2], [2, 7]]],
             ],
             np.int8,
         )
-        t = np.array([[[5, 3], [2, 7]], [[9, 8], [1, 7]]], np.int8)
+        sim = np.array([[[5, 3], [2, 7]], [[9, 8], [1, 7]]], np.int8)
 
-        assert isinstance(metric(p, t), np.ndarray)
-        assert isinstance(metric(da.from_array(p), da.from_array(t)), da.Array)
-        assert isinstance(metric(p, da.from_array(t)), da.Array)
-        assert isinstance(metric(da.from_array(p), t), da.Array)
+        assert isinstance(metric(expt, sim), np.ndarray)
+        assert isinstance(
+            metric(da.from_array(expt), da.from_array(sim)), da.Array
+        )
+        assert isinstance(metric(expt, da.from_array(sim)), da.Array)
+        assert isinstance(metric(da.from_array(expt), sim), da.Array)
 
     def test_flat_metric(self):
-        p = np.array(
+        expt = np.array(
             [
                 [[[1, 2], [3, 4]], [[5, 6], [7, 8]]],
                 [[[9, 8], [1, 7]], [[5, 2], [2, 7]]],
             ],
             np.int8,
         )
-        t = np.array([[[5, 3], [2, 7]], [[9, 8], [1, 7]]], np.int8)
+        sim = np.array([[[5, 3], [2, 7]], [[9, 8], [1, 7]]], np.int8)
         euclidean_metric = make_similarity_metric(
-            lambda p, t: cdist(p, t, metric="euclidean"),
+            lambda expt, sim: cdist(expt, sim, metric="euclidean"),
             greater_is_better=False,
             flat=True,
             scope=MetricScope.MANY_TO_MANY,
             make_compatible_to_lower_scopes=True,
         )
         assert (
-            euclidean_metric._is_compatible(p.ndim, t.ndim) is True
-            and pytest.approx(euclidean_metric(p, t)[2, 1]) == 0
+            euclidean_metric._is_compatible(expt.ndim, sim.ndim) is True
+            and pytest.approx(euclidean_metric(expt, sim)[2, 1]) == 0
         )
 
     def test_make_compatible_to_lower_scopes(self):
@@ -137,7 +145,7 @@ class TestSimilarityMetrics:
 
     def test_too_large_scoped_inputs(self):
         metric = make_similarity_metric(
-            lambda p, t: 1.0, scope=MetricScope.ONE_TO_ONE
+            lambda expt, sim: 1.0, scope=MetricScope.ONE_TO_ONE
         )
         assert (
             metric._is_compatible(
@@ -148,7 +156,7 @@ class TestSimilarityMetrics:
 
     def test_not_supported_inputs(self):
         metric = make_similarity_metric(
-            lambda p, t: 1.0,
+            lambda expt, sim: 1.0,
             scope=MetricScope.MANY_TO_MANY,
             make_compatible_to_lower_scopes=True,
         )
@@ -161,28 +169,29 @@ class TestSimilarityMetrics:
 
     def test_too_small_scoped_inputs(self):
         metric = make_similarity_metric(
-            lambda p, t: np.zeros((2, 2, 2)), scope=MetricScope.MANY_TO_MANY
+            lambda expt, sim: np.zeros((2, 2, 2)),
+            scope=MetricScope.MANY_TO_MANY,
         )
         assert (
             metric._is_compatible(np.zeros((2, 2)).ndim, np.zeros((2, 2)).ndim)
             is False
         )
 
-    def test_get_number_of_templates(self):
-        t = np.array([[[5, 3], [2, 7]], [[9, 8], [1, 7]]], np.int8)
+    def test_get_number_of_simulated(self):
+        sim = np.array([[[5, 3], [2, 7]], [[9, 8], [1, 7]]], np.int8)
         assert (
-            _get_number_of_templates(t) == 2
-            and _get_number_of_templates(t[0]) == 1
+            _get_number_of_simulated(sim) == 2
+            and _get_number_of_simulated(sim[0]) == 1
         )
 
     def test_similarity_metric_representation(self):
         metrics = [
             make_similarity_metric(
-                metric_func=lambda p, t: np.zeros((2, 2, 2)),
+                metric_func=lambda expt, sim: np.zeros((2, 2, 2)),
                 scope=MetricScope.MANY_TO_MANY,
             ),
             make_similarity_metric(
-                metric_func=lambda p, t: np.zeros((2, 2, 2)),
+                metric_func=lambda expt, sim: np.zeros((2, 2, 2)),
                 scope=MetricScope.ONE_TO_MANY,
                 flat=True,
             ),